Rename package test_data
