+++ src/test/java/test_data/ITestData.java
@@ -1,4 +1,4 @@
-package testData;
+package test_data;
 
 import org.testng.annotations.DataProvider;
 import utils.PropertyReader;
